+++ cmd/template/cluster/runner.go
@@ -67,7 +67,6 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 			MasterAZ:       r.flag.MasterAZ,
 			Description:    r.flag.Name,
 			Owner:          r.flag.Owner,
-			Region:         r.flag.Region,
 			ReleaseVersion: r.flag.Release,
 			Namespace:      metav1.NamespaceDefault,
 		}